requirements.txt
torch>=2.2.0
gymnasium>=0.29.0
stable-baselines3>=2.3.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
pandas>=2.2.0
pyarrow>=15.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pytest>=8.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
websockets>=12.0
